added the ability to provide multiple options to Study.data method
